Move update_tracker directly under hallmonitor module
